replicator/requirements.txt
dnspython~=1.16.0
requests~=2.21.0
